move: 예약 상태 enum 파일 entity 폴더에서 vo 폴더로 이동
