+++ store-reservation-service/src/main/java/com/monkey/storereservationservice/application/dto/response/ResStoreReservationGetDTOApiV1.java
@@ -1,6 +1,6 @@
 package com.monkey.storereservationservice.application.dto.response;
 
-import com.monkey.storereservationservice.domain.StoreReservation.entity.StoreReservationStatus;
+import com.monkey.storereservationservice.domain.StoreReservation.vo.StoreReservationStatus;
 import lombok.Builder;
 import lombok.Getter;
