+++ store-reservation-service/src/main/java/com/monkey/storereservationservice/domain/StoreReservation/vo/StoreReservationStatus.java
@@ -1,4 +1,4 @@
-package com.monkey.storereservationservice.domain.StoreReservation.entity;
+package com.monkey.storereservationservice.domain.StoreReservation.vo;
 
 public enum StoreReservationStatus {
     SCHEDULED,
